requirements.txt
streamlit
psycopg2
sqlalchemy
pandas
tiktoken
openai
python-dotenv
asyncio
